Revert "More datasets (#64)"
This reverts commit 5e4c4e911d37d53342ad38b5f947b8e1ec9b41a7.
