+++ crates/cli/src/args.rs
@@ -155,37 +155,9 @@ pub struct Args {
     #[arg(long, help_heading = "Output Options")]
     pub no_report: bool,
 
-    /// Address
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..))]
-    pub address: Option<Vec<String>>,
-
-    /// To Address
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..), value_name="TO")]
-    pub to_address: Option<Vec<String>>,
-
-    /// From Address
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..), value_name="FROM")]
-    pub from_address: Option<Vec<String>>,
-
-    /// [eth_calls] Call data to use for eth_calls
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..))]
-    pub call_data: Option<Vec<String>>,
-
-    /// [eth_calls] Function to use for eth_calls
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..))]
-    pub function: Option<Vec<String>>,
-
-    /// [eth_calls] Inputs to use for eth_calls
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..))]
-    pub inputs: Option<Vec<String>>,
-
-    /// [slots] Slots
-    #[arg(long, help_heading = "Dataset-specific Options", num_args(1..))]
-    pub slots: Option<Vec<String>>,
-
     /// [logs] filter logs by contract address
     #[arg(long, help_heading = "Dataset-specific Options")]
-    pub contract: Option<Vec<String>>,
+    pub contract: Option<String>,
 
     /// [logs] filter logs by topic0
     #[arg(long, visible_alias = "event", help_heading = "Dataset-specific Options")]
@@ -206,14 +178,14 @@ pub struct Args {
     /// [logs] Blocks per request
     #[arg(
         long,
-        value_name = "SIZE",
+        value_name = "BLOCKS",
         default_value_t = 1,
         help_heading = "Dataset-specific Options"
     )]
     pub inner_request_size: u64,
 
     /// [logs] event signature to parse
-    #[arg(long, value_name = "SIGNATURE", help_heading = "Dataset-specific Options")]
+    #[arg(long, help_heading = "Dataset-specific Options")]
     pub event_signature: Option<String>,
 }
