+++ crates/cli/src/parse/mod.rs
@@ -1,11 +1,12 @@
 mod args;
 mod blocks;
 mod file_output;
-mod parse_utils;
 mod query;
-mod schemas;
 mod source;
 mod transactions;
 
 pub use args::*;
-use schemas::*;
+// use blocks::*;
+// use file_output::*;
+// use query::*;
+// use source::*;
