+++ crates/cli/src/parse/query.rs
@@ -1,141 +1,233 @@
-use std::{collections::HashMap, sync::Arc};
+use std::{
+    collections::{HashMap, HashSet},
+    sync::Arc,
+};
 
 use ethers::prelude::*;
 use hex::FromHex;
 
 use cryo_freeze::{
-    AddressChunk, CallDataChunk, Datatype, Fetcher, MultiQuery, ParseError, RowFilter, SlotChunk,
+    ColumnEncoding, Datatype, Fetcher, FileFormat, LogDecoder, MultiQuery, ParseError, RowFilter,
+    Table,
 };
 
-use super::{
-    blocks, parse_schemas,
-    parse_utils::{hex_string_to_binary, hex_strings_to_binary, parse_binary_arg},
-    transactions,
-};
+use super::{blocks, file_output, transactions};
 use crate::args::Args;
+use cryo_freeze::U256Type;
 
 pub(crate) async fn parse_query<P: JsonRpcClient>(
     args: &Args,
     fetcher: Arc<Fetcher<P>>,
 ) -> Result<MultiQuery, ParseError> {
-    // process schemas
-    let schemas = parse_schemas(args)?;
-
-    // process chunks
     let chunks = match (&args.blocks, &args.txs) {
         (Some(_), None) => blocks::parse_blocks(args, fetcher).await?,
         (None, Some(txs)) => transactions::parse_transactions(txs)?,
-        (None, None) => blocks::get_default_block_chunks(args, fetcher, &schemas).await?,
+        (None, None) => blocks::get_default_block_chunks(args, fetcher).await?,
         (Some(_), Some(_)) => {
             return Err(ParseError::ParseError("specify only one of --blocks or --txs".to_string()))
         }
     };
 
-    // deprecated
-    let address = if let Some(contract) = &args.contract {
-        parse_address(&Some(contract[0].clone()))
-    } else {
-        None
-    };
+    // process schemas
+    let schemas = parse_schemas(args)?;
 
     // build row filters
-    let call_data_chunks = parse_call_datas(&args.call_data, &args.function, &args.inputs)?;
-    let address_chunks = parse_address_chunks(&args.address, "address")?;
-    let contract_chunks = parse_address_chunks(&args.contract, "contract_address")?;
-    let to_address_chunks = parse_address_chunks(&args.to_address, "to_address")?;
-    let slot_chunks = parse_slot_chunks(&args.slots, "slot")?;
+    let contract = parse_address(&args.contract);
     let topics = [
         parse_topic(&args.topic0),
         parse_topic(&args.topic1),
         parse_topic(&args.topic2),
         parse_topic(&args.topic3),
     ];
-    let row_filter = RowFilter {
-        address,
-        topics,
-        address_chunks,
-        contract_chunks,
-        to_address_chunks,
-        slot_chunks,
-        call_data_chunks,
-    };
+    let row_filter = RowFilter { address: contract, topics };
     let mut row_filters: HashMap<Datatype, RowFilter> = HashMap::new();
-    for datatype in schemas.keys() {
-        let row_filter = row_filter.apply_arg_aliases(datatype.dataset().arg_aliases());
-        row_filters.insert(*datatype, row_filter);
-    }
+    row_filters.insert(Datatype::Logs, row_filter);
 
     let query = MultiQuery { schemas, chunks, row_filters };
     Ok(query)
 }
 
-fn parse_call_datas(
-    call_datas: &Option<Vec<String>>,
-    function: &Option<Vec<String>>,
-    inputs: &Option<Vec<String>>,
-) -> Result<Option<Vec<CallDataChunk>>, ParseError> {
-    let call_datas = match (call_datas, function, inputs) {
-        (None, None, None) => return Ok(None),
-        (Some(call_data), None, None) => hex_strings_to_binary(call_data)?,
-        (None, Some(function), None) => hex_strings_to_binary(function)?,
-        (None, Some(function), Some(inputs)) => {
-            let mut call_datas = Vec::new();
-            for f in function.iter() {
-                for i in inputs.iter() {
-                    let mut call_data = hex_string_to_binary(f)?.clone();
-                    call_data.extend(hex_string_to_binary(i)?);
-                    call_datas.push(call_data);
+fn parse_datatypes(raw_inputs: &Vec<String>) -> Result<Vec<Datatype>, ParseError> {
+    let mut datatypes = Vec::new();
+
+    for raw_input in raw_inputs {
+        match raw_input.as_str() {
+            "state_diffs" => {
+                datatypes.push(Datatype::BalanceDiffs);
+                datatypes.push(Datatype::CodeDiffs);
+                datatypes.push(Datatype::NonceDiffs);
+                datatypes.push(Datatype::StorageDiffs);
+            }
+            datatype => {
+                let datatype = match datatype {
+                    "balance_diffs" => Datatype::BalanceDiffs,
+                    "blocks" => Datatype::Blocks,
+                    "code_diffs" => Datatype::CodeDiffs,
+                    "logs" => Datatype::Logs,
+                    "events" => Datatype::Logs,
+                    "nonce_diffs" => Datatype::NonceDiffs,
+                    "storage_diffs" => Datatype::StorageDiffs,
+                    "transactions" => Datatype::Transactions,
+                    "txs" => Datatype::Transactions,
+                    "traces" => Datatype::Traces,
+                    "vm_traces" => Datatype::VmTraces,
+                    "opcode_traces" => Datatype::VmTraces,
+                    "native_transfers" => Datatype::NativeTransfers,
+                    "contracts" => Datatype::Contracts,
+                    _ => {
+                        return Err(ParseError::ParseError(format!("invalid datatype {}", datatype)))
+                    }
+                };
+                datatypes.push(datatype)
+            }
+        }
+    }
+    Ok(datatypes)
+}
+
+fn parse_schemas(args: &Args) -> Result<HashMap<Datatype, Table>, ParseError> {
+    let datatypes = parse_datatypes(&args.datatype)?;
+    let output_format = file_output::parse_output_format(args)?;
+
+    let u256_types = if let Some(raw_u256_types) = &args.u256_types {
+        let mut u256_types: HashSet<U256Type> = HashSet::new();
+        for raw in raw_u256_types.iter() {
+            let u256_type = match raw.to_lowercase() {
+                raw if raw == "binary" => U256Type::Binary,
+                raw if raw == "string" => U256Type::String,
+                raw if raw == "str" => U256Type::String,
+                raw if raw == "f32" => U256Type::F32,
+                raw if raw == "float32" => U256Type::F32,
+                raw if raw == "f64" => U256Type::F64,
+                raw if raw == "float64" => U256Type::F64,
+                raw if raw == "float" => U256Type::F64,
+                raw if raw == "u32" => U256Type::U32,
+                raw if raw == "uint32" => U256Type::U32,
+                raw if raw == "u64" => U256Type::U64,
+                raw if raw == "uint64" => U256Type::U64,
+                raw if raw == "decimal128" => U256Type::Decimal128,
+                raw if raw == "d128" => U256Type::Decimal128,
+                _ => return Err(ParseError::ParseError("bad u256 type".to_string())),
+            };
+            u256_types.insert(u256_type);
+        }
+        u256_types
+    } else {
+        HashSet::from_iter(vec![U256Type::Binary, U256Type::String, U256Type::F64])
+    };
+    let binary_column_format = match args.hex | (output_format != FileFormat::Parquet) {
+        true => ColumnEncoding::Hex,
+        false => ColumnEncoding::Binary,
+    };
+
+    let log_decoder = match args.event_signature {
+        Some(ref sig) => LogDecoder::new(sig.clone()).ok(),
+        None => None,
+    };
+
+    let sort = parse_sort(&args.sort, &datatypes)?;
+    let schemas: Result<HashMap<Datatype, Table>, ParseError> = datatypes
+        .iter()
+        .map(|datatype| {
+            datatype
+                .table_schema(
+                    &u256_types,
+                    &binary_column_format,
+                    &args.include_columns,
+                    &args.exclude_columns,
+                    &args.columns,
+                    sort[datatype].clone(),
+                    log_decoder.clone(),
+                )
+                .map(|schema| (*datatype, schema))
+                .map_err(|_e| {
+                    ParseError::ParseError(format!(
+                        "Failed to get schema for datatype: {:?}",
+                        datatype
+                    ))
+                })
+        })
+        .collect();
+
+    // make sure all included columns ended up in at least one schema
+    if let (Ok(schemas), Some(include_columns)) = (&schemas, &args.include_columns) {
+        let mut unknown_columns = Vec::new();
+        for column in include_columns.iter() {
+            let mut in_a_schema = false;
+
+            for schema in schemas.values() {
+                if schema.has_column(column) {
+                    in_a_schema = true;
+                    break
                 }
             }
-            call_datas
+
+            if !in_a_schema {
+                unknown_columns.push(column);
+            }
         }
-        (None, None, Some(_)) => {
-            let message = "must specify function if specifying inputs";
-            return Err(ParseError::ParseError(message.to_string()))
-        }
-        (Some(_), Some(_), None) => {
-            let message = "cannot specify both call_data and function";
-            return Err(ParseError::ParseError(message.to_string()))
-        }
-        (Some(_), None, Some(_)) => {
-            let message = "cannot specify both call_data and inputs";
-            return Err(ParseError::ParseError(message.to_string()))
-        }
-        (Some(_), Some(_), Some(_)) => {
-            let message = "cannot specify both call_data and function";
-            return Err(ParseError::ParseError(message.to_string()))
+        if !unknown_columns.is_empty() {
+            return Err(ParseError::ParseError(format!(
+                "datatypes do not support these columns: {:?}",
+                unknown_columns
+            )))
         }
     };
-    Ok(Some(vec![CallDataChunk::Values(call_datas)]))
+
+    // make sure all excluded columns are excluded from at least one schema
+    if let (Ok(schemas), Some(exclude_columns)) = (&schemas, &args.exclude_columns) {
+        let mut unknown_columns = Vec::new();
+        for column in exclude_columns.iter() {
+            let mut in_a_schema = false;
+
+            for datatype in schemas.keys() {
+                if datatype.dataset().column_types().contains_key(&column.as_str()) {
+                    in_a_schema = true;
+                    break
+                }
+            }
+
+            if !in_a_schema {
+                unknown_columns.push(column);
+            }
+        }
+        if !unknown_columns.is_empty() {
+            return Err(ParseError::ParseError(format!(
+                "datatypes do not support these columns: {:?}",
+                unknown_columns
+            )))
+        }
+    };
+
+    schemas
 }
 
-pub(crate) fn parse_address_chunks(
-    address: &Option<Vec<String>>,
-    default_column: &str,
-) -> Result<Option<Vec<AddressChunk>>, ParseError> {
-    if let Some(address) = address {
-        let chunks = parse_binary_arg(address, default_column)?
-            .values()
-            .map(|a| AddressChunk::Values(a.clone()))
-            .collect();
-        Ok(Some(chunks))
-    } else {
-        Ok(None)
-    }
-}
-
-pub(crate) fn parse_slot_chunks(
-    slots: &Option<Vec<String>>,
-    default_column: &str,
-) -> Result<Option<Vec<SlotChunk>>, ParseError> {
-    if let Some(values) = slots {
-        let chunks = parse_binary_arg(values, default_column)?
-            .values()
-            .map(|a| SlotChunk::Values(a.clone()))
-            .collect();
-        Ok(Some(chunks))
-    } else {
-        Ok(None)
+fn parse_sort(
+    raw_sort: &Option<Vec<String>>,
+    datatypes: &Vec<Datatype>,
+) -> Result<HashMap<Datatype, Option<Vec<String>>>, ParseError> {
+    match raw_sort {
+        None => Ok(HashMap::from_iter(
+            datatypes.iter().map(|datatype| (*datatype, Some(datatype.dataset().default_sort()))),
+        )),
+        Some(raw_sort) => {
+            if (raw_sort.len() == 1) && (raw_sort[0] == "none") {
+                Ok(HashMap::from_iter(datatypes.iter().map(|datatype| (*datatype, None))))
+            } else if raw_sort.is_empty() {
+                Err(ParseError::ParseError(
+                    "must specify columns to sort by, use `none` to disable sorting".to_string(),
+                ))
+            } else if datatypes.len() > 1 {
+                Err(ParseError::ParseError(
+                    "custom sort not supported for multiple datasets".to_string(),
+                ))
+            } else {
+                match datatypes.iter().next() {
+                    Some(datatype) => Ok(HashMap::from_iter([(*datatype, Some(raw_sort.clone()))])),
+                    None => Err(ParseError::ParseError("schemas map is empty".to_string())),
+                }
+            }
+        }
     }
 }
