+++ crates/freeze/Cargo.toml
@@ -17,7 +17,6 @@ futures = { workspace = true }
 governor = { workspace = true }
 indexmap = { workspace = true }
 indicatif = { workspace = true }
-lazy_static = { workspace = true }
 polars = { workspace = true }
 prefix-hex = { workspace = true }
 serde = { workspace = true }
