+++ crates/freeze/src/datasets/mod.rs
@@ -1,27 +1,14 @@
 mod balance_diffs;
-mod balances;
 mod blocks;
 mod blocks_and_transactions;
 mod code_diffs;
-mod codes;
 mod contracts;
-mod erc20_balances;
-mod erc20_metadata;
-mod erc20_supplies;
-mod erc20_transfers;
-mod erc721_metadata;
-mod erc721_transfers;
-mod eth_calls;
 mod logs;
 mod native_transfers;
 mod nonce_diffs;
-mod nonces;
 mod state_diffs;
 mod storage_diffs;
-mod storages;
-mod trace_calls;
 mod traces;
-mod transaction_addresses;
 mod transactions;
 mod vm_traces;
