+++ crates/freeze/src/datasets/traces.rs
@@ -166,7 +166,7 @@ pub(crate) fn fetch_transaction_traces(
     }
 }
 
-pub(crate) fn reward_type_to_string(reward_type: &RewardType) -> String {
+fn reward_type_to_string(reward_type: &RewardType) -> String {
     match reward_type {
         RewardType::Block => "reward".to_string(),
         RewardType::Uncle => "uncle".to_string(),
@@ -175,7 +175,7 @@ pub(crate) fn reward_type_to_string(reward_type: &RewardType) -> String {
     }
 }
 
-pub(crate) fn action_type_to_string(action_type: &ActionType) -> String {
+fn action_type_to_string(action_type: &ActionType) -> String {
     match action_type {
         ActionType::Call => "call".to_string(),
         ActionType::Create => "create".to_string(),
@@ -184,7 +184,7 @@ pub(crate) fn action_type_to_string(action_type: &ActionType) -> String {
     }
 }
 
-pub(crate) fn action_call_type_to_string(action_call_type: &CallType) -> String {
+fn action_call_type_to_string(action_call_type: &CallType) -> String {
     match action_call_type {
         CallType::None => "none".to_string(),
         CallType::Call => "call".to_string(),
