+++ crates/freeze/src/types/chunks/binary_chunk.rs
@@ -44,13 +44,3 @@ impl ChunkData for BinaryChunk {
         }
     }
 }
-
-impl BinaryChunk {
-    /// get list of values in chunk
-    pub fn values(&self) -> &Vec<Vec<u8>> {
-        match self {
-            BinaryChunk::Values(values) => values,
-            BinaryChunk::Range(_start, _end) => panic!("values not implemented for binary ranges"),
-        }
-    }
-}
