+++ crates/freeze/src/types/chunks/mod.rs
@@ -4,6 +4,6 @@ pub(crate) mod chunk_ops;
 pub(crate) mod number_chunk;
 pub(crate) mod subchunks;
 
-pub use chunk::{AddressChunk, BlockChunk, CallDataChunk, Chunk, SlotChunk, TransactionChunk};
+pub use chunk::{AddressChunk, BlockChunk, Chunk, TransactionChunk};
 pub use chunk_ops::ChunkData;
 pub use subchunks::Subchunk;
