+++ crates/freeze/src/types/dataframes/mod.rs
@@ -1,10 +1,8 @@
 mod export;
-mod read;
 mod sort;
 
 #[macro_use]
 mod creation;
 
 pub(crate) use export::*;
-pub use read::*;
 pub(crate) use sort::SortableDataFrame;
