+++ crates/freeze/src/types/queries.rs
@@ -7,7 +7,7 @@ use ethers::prelude::*;
 
 use crate::{
     types::{Chunk, Datatype, Table},
-    AddressChunk, CallDataChunk, CollectError, FileOutput, FreezeError, SlotChunk,
+    CollectError, FileOutput, FreezeError,
 };
 
 /// Query multiple data types
@@ -69,124 +69,12 @@ fn list_files(dir: &Path) -> Result<Vec<PathBuf>, std::io::Error> {
 }
 
 /// Options for fetching logs
-#[derive(Clone, Default)]
+#[derive(Clone)]
 pub struct RowFilter {
     /// topics to filter for
     pub topics: [Option<ValueOrArray<Option<H256>>>; 4],
     /// address to filter for
     pub address: Option<ValueOrArray<H160>>,
-    /// address chunks to collect
-    pub address_chunks: Option<Vec<AddressChunk>>,
-    /// contract chunks to collect
-    pub contract_chunks: Option<Vec<AddressChunk>>,
-    /// to_address chunks to collect
-    pub to_address_chunks: Option<Vec<AddressChunk>>,
-    /// slot chunks to collect
-    pub slot_chunks: Option<Vec<SlotChunk>>,
-    /// call_data chunks to collect
-    pub call_data_chunks: Option<Vec<CallDataChunk>>,
-}
-
-impl RowFilter {
-    pub(crate) fn address_chunks(&self) -> Result<Vec<AddressChunk>, CollectError> {
-        match &self.address_chunks {
-            Some(address_chunks) => Ok(address_chunks.clone()),
-            _ => Err(CollectError::CollectError("must specify --address".to_string())),
-        }
-    }
-
-    pub(crate) fn contract_chunks(&self) -> Result<Vec<AddressChunk>, CollectError> {
-        match &self.contract_chunks {
-            Some(contract_chunks) => Ok(contract_chunks.clone()),
-            _ => Err(CollectError::CollectError("must specify --contract".to_string())),
-        }
-    }
-
-    pub(crate) fn to_address_chunks(&self) -> Result<Vec<AddressChunk>, CollectError> {
-        match &self.to_address_chunks {
-            Some(to_address_chunks) => Ok(to_address_chunks.clone()),
-            _ => Err(CollectError::CollectError("must specify --to-address".to_string())),
-        }
-    }
-
-    // pub(crate) fn from_address_chunks(&self) -> Result<Vec<AddressChunk>, CollectError> {
-    //     match &self.from_address_chunks {
-    //         Some(from_address_chunks) => Ok(from_address_chunks.clone()),
-    //         _ => Err(CollectError::CollectError("must specify --from-address".to_string())),
-    //     }
-    // }
-
-    pub(crate) fn slot_chunks(&self) -> Result<Vec<SlotChunk>, CollectError> {
-        match &self.slot_chunks {
-            Some(slot_chunks) => Ok(slot_chunks.clone()),
-            _ => Err(CollectError::CollectError("must specify slots".to_string())),
-        }
-    }
-
-    pub(crate) fn call_data_chunks(&self) -> Result<Vec<CallDataChunk>, CollectError> {
-        match &self.call_data_chunks {
-            Some(call_data_chunks) => Ok(call_data_chunks.clone()),
-            _ => Err(CollectError::CollectError("must specify call_data".to_string())),
-        }
-    }
-
-    /// apply arg aliases
-    pub fn apply_arg_aliases(&self, arg_aliases: HashMap<String, String>) -> RowFilter {
-        let mut row_filter: RowFilter = self.clone();
-        for (from, to) in arg_aliases.iter() {
-            row_filter = match from.as_str() {
-                "address" => match to.as_str() {
-                    "contract" => RowFilter {
-                        contract_chunks: row_filter.address_chunks.clone(),
-                        address_chunks: None,
-                        ..row_filter.clone()
-                    },
-                    "to_address" => RowFilter {
-                        to_address_chunks: row_filter.address_chunks.clone(),
-                        address_chunks: None,
-                        ..row_filter.clone()
-                    },
-                    _ => {
-                        panic!("invalid alias")
-                    }
-                },
-                "contract" => match to.as_str() {
-                    "address" => RowFilter {
-                        address_chunks: row_filter.address_chunks.clone(),
-                        contract_chunks: None,
-                        ..row_filter.clone()
-                    },
-                    "to_address" => RowFilter {
-                        to_address_chunks: row_filter.contract_chunks.clone(),
-                        contract_chunks: None,
-                        ..row_filter.clone()
-                    },
-                    _ => {
-                        panic!("invalid alias")
-                    }
-                },
-                "to_address" => match to.as_str() {
-                    "address" => RowFilter {
-                        address_chunks: row_filter.to_address_chunks.clone(),
-                        to_address_chunks: None,
-                        ..row_filter.clone()
-                    },
-                    "contract" => RowFilter {
-                        contract_chunks: row_filter.to_address_chunks.clone(),
-                        to_address_chunks: None,
-                        ..row_filter.clone()
-                    },
-                    _ => {
-                        panic!("invalid alias")
-                    }
-                },
-                _ => {
-                    panic!("invalid alias")
-                }
-            };
-        }
-        row_filter
-    }
 }
 
 impl From<MultiQuery> for SingleQuery {
