+++ crates/python/src/collect_adapter.rs
@@ -40,13 +40,6 @@ use cryo_freeze::collect;
         compression = vec!["lz4".to_string()],
         report_dir = None,
         no_report = false,
-        address = None,
-        to_address = None,
-        from_address = None,
-        call_data = None,
-        function = None,
-        inputs = None,
-        slots = None,
         contract = None,
         topic0 = None,
         topic1 = None,
@@ -92,14 +85,7 @@ pub fn _collect(
     compression: Vec<String>,
     report_dir: Option<String>,
     no_report: bool,
-    address: Option<Vec<String>>,
-    to_address: Option<Vec<String>>,
-    from_address: Option<Vec<String>>,
-    call_data: Option<Vec<String>>,
-    function: Option<Vec<String>>,
-    inputs: Option<Vec<String>>,
-    slots: Option<Vec<String>>,
-    contract: Option<Vec<String>>,
+    contract: Option<String>,
     topic0: Option<String>,
     topic1: Option<String>,
     topic2: Option<String>,
@@ -141,13 +127,6 @@ pub fn _collect(
         compression,
         report_dir,
         no_report,
-        address,
-        to_address,
-        from_address,
-        call_data,
-        function,
-        inputs,
-        slots,
         contract,
         topic0,
         topic1,
