+++ crates/python/src/freeze_adapter.rs
@@ -41,13 +41,6 @@ use cryo_cli::{run, Args};
         compression = vec!["lz4".to_string()],
         report_dir = None,
         no_report = false,
-        address = None,
-        to_address = None,
-        from_address = None,
-        call_data = None,
-        function = None,
-        inputs = None,
-        slots = None,
         contract = None,
         topic0 = None,
         topic1 = None,
@@ -93,14 +86,7 @@ pub fn _freeze(
     compression: Vec<String>,
     report_dir: Option<String>,
     no_report: bool,
-    address: Option<Vec<String>>,
-    to_address: Option<Vec<String>>,
-    from_address: Option<Vec<String>>,
-    call_data: Option<Vec<String>>,
-    function: Option<Vec<String>>,
-    inputs: Option<Vec<String>>,
-    slots: Option<Vec<String>>,
-    contract: Option<Vec<String>>,
+    contract: Option<String>,
     topic0: Option<String>,
     topic1: Option<String>,
     topic2: Option<String>,
@@ -142,13 +128,6 @@ pub fn _freeze(
         compression,
         report_dir,
         no_report,
-        address,
-        to_address,
-        from_address,
-        call_data,
-        function,
-        inputs,
-        slots,
         contract,
         topic0,
         topic1,
